Allow applying agent to the daemons spawned from "embedded" runner
The crucial part is to make the agent's JAR part of the application
(aka system) classpath; we make it a runtime dependency of the test
to achieve that. This doesn't actually "apply" the agent, just makes its
classes accessible, and allows the ModuleRegistry built out of the
EffectiveClassPath to locate the agent JAR.

Note that this doesn't apply the agent to the test processes, so agent
isn't used if the embedded runner runs the build.