Allow executing builds with legacy instrumentation with "--no-daemon"
The agent is applied to the client process in the start script where it
is hard to check for the feature flag. However, if the user doesn't
enable the feature flag, the client process would become incompatible
and require the single-use daemon to be spawned. To avoid this, this
commit makes process with agent compatible with the legacy
instrumentation mode.