Handle instrumentation agent in foreground daemon mode
The foreground daemon infers its options from the JVM command line and
thus sees the "-javaagent" switch. When this switch becomes part of the
DaemonContext, all clients become incompatible, because the client
doesn't add the javaagent switch to the daemon context it expects.

It is very tempting to add the agent switch to jvm opts of the client
too, but this is a lot less trivial. It is also possible to apply the
agent without using a switch, though this isn't used anywhere (yet).

The foreground mode isn't a very popular feature, and is already tricky
to set up in Java 9+, so containing ugly bits here seems to be a good
tradeoff.